requirements.txt
streamlit==0.87.0
PyMuPDF==1.18.16
python-pptx==0.6.19